+++ modules/landing_zone/scripts/libs.py
@@ -7,3 +7,16 @@ def cli(args, root_path):
         return (1, error)
     else:
         return (0, result)
+
+def execWithErrors(args_list):
+    for args in args_list:
+        (error, result) = cli(args, os.environ['root'])
+        if error != 0:
+            print("Error: failed to execute command:")
+            raise Exception(result)
+
+def execWithoutErrors(args_list):
+    (error, result) = cli(args_list, os.environ['root'])
+    if error != 0:
+        print("Error: failed to execute command:")
+        print(result)
